# src/problem5/src/interface/http/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Support Ticket Management API
  description: >
    Problem 5 — Support Ticket Management API. Describes the approved API
    contract defined in context/api-contract.md.
  version: "1.0.0"
servers:
  - url: /api/v1
paths:
  /tickets:
    post:
      summary: Create a Ticket
      operationId: createTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketResponse"
        "400":
          description: Request body is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      summary: List Tickets
      operationId: listTickets
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TicketStatus"
        - name: priority
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TicketPriority"
      responses:
        "200":
          description: A collection of Tickets, optionally filtered by status and/or priority (AND semantics)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketCollectionResponse"
        "400":
          description: A query parameter contains an invalid value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tickets/{id}:
    get:
      summary: Get a Ticket
      operationId: getTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      responses:
        "200":
          description: The requested Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketResponse"
        "404":
          description: Ticket does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update a Ticket
      operationId: updateTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketRequest"
      responses:
        "200":
          description: The updated Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketResponse"
        "400":
          description: Update data is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Ticket does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a Ticket
      operationId: deleteTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      responses:
        "204":
          description: Ticket deleted — no response body
        "404":
          description: Ticket does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    TicketId:
      name: id
      in: path
      required: true
      description: Ticket identifier
      schema:
        type: string
  schemas:
    TicketStatus:
      type: string
      enum: [open, in_progress, resolved, closed]
    TicketPriority:
      type: string
      enum: [low, medium, high]
    Ticket:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/TicketStatus"
        priority:
          $ref: "#/components/schemas/TicketPriority"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        [id, title, description, status, priority, createdAt, updatedAt]
    TicketResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Ticket"
      required: [data]
    TicketCollectionResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
      required: [data]
    CreateTicketRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        priority:
          $ref: "#/components/schemas/TicketPriority"
      required: [title, description, priority]
    UpdateTicketRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        status:
          $ref: "#/components/schemas/TicketStatus"
        priority:
          $ref: "#/components/schemas/TicketPriority"
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                [
                  VALIDATION_ERROR,
                  TICKET_NOT_FOUND,
                  INTERNAL_SERVER_ERROR,
                  NOT_FOUND,
                ]
            message:
              type: string
          required: [code, message]
      required: [error]